Add proper authorization method to Restaurant Endpoints

+++ RestaurantAPI/Services/RestaurantSerivce.cs
@@ -1,18 +1,21 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
     public interface IRestaurantSerivce
     {
-        int Create(CreateRestaurantDto dto);
+        int Create(CreateRestaurantDto dto, int userId);
         IEnumerable<RestaurantDto> GetAll();
         RestaurantDto GetById(int id);
-        void Delete(int id);
-        void Modify(int id, ModifyRestaurantDto dto);
+        void Delete(int id, ClaimsPrincipal user);
+        void Modify(int id, ModifyRestaurantDto dto, ClaimsPrincipal user);
     }
 
     public class RestaurantSerivce : IRestaurantSerivce
@@ -20,14 +23,16 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantSerivce> _logger;
-        public RestaurantSerivce(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantSerivce> logger)
+        private readonly IAuthorizationService _authorizationService;
+        public RestaurantSerivce(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantSerivce> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public void Modify(int id, ModifyRestaurantDto dto)
+        public void Modify(int id, ModifyRestaurantDto dto, ClaimsPrincipal user)
         {
             var restaurant = _dbContext
                 .Restaurants
@@ -36,7 +41,14 @@ namespace RestaurantAPI.Services
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found.");
 
-            if(dto.Name != null)
+            var authorizationResulte = _authorizationService.AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Modify)).Result;
+
+            if (!authorizationResulte.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
+            if (dto.Name != null)
             {
                 restaurant.Name = dto.Name;
             }
@@ -53,7 +65,7 @@ namespace RestaurantAPI.Services
             _dbContext.SaveChanges();
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Restaurant with id: {id} DELETE action invoked.");
 
@@ -63,6 +75,13 @@ namespace RestaurantAPI.Services
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found.");
+
+            var authorizationResulte = _authorizationService.AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResulte.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
@@ -108,9 +127,10 @@ namespace RestaurantAPI.Services
             return restaurantsDtos;
         }
 
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto, int userId)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
